at91: drop target
This target is still on kernel 4.14, and no attempt has been made
to update it. Since we already are two LTS versions ahead of that,
the target is dropped, as the chance of somebody bumping it will
only decrease with time.

This also drops the target-specific packages.

Note:
kmod-usb-ohci still selects CONFIG_USB_OHCI_HCD_AT91 and ohci-at91.ko
I'm not sure whether those are required elsewhere, so I didn't touch
them.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>