libtool: fix hard-coded paths
libtoolize hardcodes some paths. This is fine when building libtool and
then using it (for example in OpenWrt's buildroot). But when using an
SDK the paths are most likely different.

For example, when building util-linux within an SDK we're greeted with
the following message:

  libtoolize:   error: $pkgauxdir is not a directory: '/builder/shared-workdir/build/staging_dir/host/share/libtool/build-aux'

This is because staging_dir/host/bin/libtoolize contains the following
hard-coded paths from when the SDK was built in the first place:

    prefix="/builder/shared-workdir/build/staging_dir/host"
    datadir="/builder/shared-workdir/build/staging_dir/host/share"
    pkgauxdir="/builder/shared-workdir/build/staging_dir/host/share/libtool/build-aux"
    pkgltdldir="/builder/shared-workdir/build/staging_dir/host/share/libtool"
    aclocaldir="/builder/shared-workdir/build/staging_dir/host/share/aclocal"

This commits updates 000-relocatable.patch to correct the paths,
relative to "$STAGING_DIR".

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>