tools: add python3 host build option
With the increasing changes to the Python platform,
such as modules being added/removed, some users
may have problems using their host machine's
version of python3, or they may only have
a beta version that is not yet supported
in the prereq-build checks before a build starts.

Add an option to build python3 from source
to help those users who may have a hard time
getting a different version on their host machine,
or for whatever reason do not want to change versions.

Having this as an option allows for more control:
there may at some point be a solid reason
to make this option default to yes,
and that reason may pass allowing
maintainers to set it back to no.

Signed-off-by: Michael Pratt <mcpratt@pm.me>