kernel: fw_devlink: some improvement patches
These patches are from a series sent upstream
by Saravana Kannan, and they likely will be merged
after a small revision.

Functions introduced here are part of a solution
to an issue in fw_devlink regarding supplier fwnodes
that do not represent a real device or any driver
and therefore cause probe defer of the consumer,
but have a parent node which does represent a real device.
(e.g. MTD partitions that are an NVMEM provider with "nvmem-cells")

These patches alone do not solve the problem in the case
where the consumer fwnode is parsed for fwnode/device links
before the supplier fwnode is (consumer fwnode is ordered
first in the devicetree descendants, so it parses and probes first).
In this case, before fwnode links can be modified during supplier probe,
they are already dropped because of device link creation by the consumer.
(handled by the patches in the next commits)

Link: https://lore.kernel.org/lkml/20220810060040.321697-1-saravanak@google.com/
Signed-off-by: Michael Pratt <mcpratt@pm.me>