kernel: fw_devlink: backport some improvement patches
These patches are from patch series sent upstream
by Saravana Kannan, some directly related to fw_devlink,
and some indirectly related in order to sync with upstream.

Functions introduced here are part of a solution
to an issue in fw_devlink regarding supplier fwnodes
that do not represent a real device or any driver
and therefore cause probe defer of the consumer,
but have a parent node which does represent a real device.
(e.g. MTD partitions that are an NVMEM provider with "nvmem-cells")

These patches alone do not solve the problem in the case
where the consumer fwnode is parsed for fwnode/device links
before the supplier fwnode is (consumer fwnode is ordered
first in the devicetree descendants, so it parses and probes first).
In this case, before fwnode links can be modified during supplier probe,
they are already dropped during device link creation by the consumer.

Link: https://lore.kernel.org/r/20220810060040.321697-1-saravanak@google.com
Link: https://lore.kernel.org/r/20230127001141.407071-1-saravanak@google.com
Link: https://lore.kernel.org/r/20230207014207.1678715-1-saravanak@google.com
Signed-off-by: Michael Pratt <mcpratt@pm.me>