generic: fw_devlink: backport some improvement patches
These patches are from multiple patch series
by Saravana Kannan at Google, some directly related to fw_devlink,
and some indirectly related in order to sync with upstream.

Functions introduced here are part of a solution
to an issue in fw_devlink regarding supplier fwnodes
that do not represent a real device or have any driver
and therefore cause probe defer of the consumer,
but have a parent node which does represent a real device.
(e.g. MTD partitions that are an NVMEM provider with "nvmem-cells")

These patches alone do not solve the problem in the case
where the consumer fwnode is parsed for fwnode/device links
before the supplier fwnode is (consumer fwnode is ordered
first in the devicetree descendants, so it parses and probes first).
In this case, before fwnode links can be modified during supplier probe,
they are already dropped during device link creation by the consumer.

The upstream commit 4a032827daa89350365166b19d14d82fe8219128
("of: property: Simplify of_link_to_phandle()") is not currently included,
which most notably removes the use of of_get_compat_node() for fw_devlink.

The main goal of the fw_devlink improvement series by Saravana is for
fw_devlink to no longer rely on the "compatible" property, however
in several cases, including "nvmem-cells", the nearest node
with a "compatible" property is a good place to start for the supplier,
as a node representing a real device is extremely likely to have one.

At the time of this commit, the upstream commit
is not backported into 5.15, but is in 6.1.53.
A deeper dive is needed to see why the simplification
at of_link_to_phandle() is causing issues, but in the meantime,
it can be safely reverted.

The value of driver_deferred_probe_timeout being non-zero
works in 6.1 but not in 5.15, and likely would require
more backport commits for 5.15, but is not necessary for devlinks,
so backport in 5.15 a temporary commit setting it to zero.

Link: https://lore.kernel.org/r/20220429220933.1350374-1-saravanak@google.com
Link: https://lore.kernel.org/r/20220526034609.480766-1-saravanak@google.com
Link: https://lore.kernel.org/r/20220810060040.321697-1-saravanak@google.com
Link: https://lore.kernel.org/r/20230127001141.407071-1-saravanak@google.com
Link: https://lore.kernel.org/r/20230207014207.1678715-1-saravanak@google.com
Signed-off-by: Michael Pratt <mcpratt@pm.me>