ramips: disable driverless devices
There is no driver available
for "ralink,rt2880-clock" or "ralink,rt3050-usbphy",
so set the status to "disabled".

Without a "disabled" status, the kernel's fw_devlink feature
creates links between these nodes and nodes with phandles to it,
but the missing driver causes infinite probing defer for both
the supplier and consumer device as a result of those links.

Fixes: ffaaa6788 ("ramips: DTS modifications")
Signed-off-by: Michael Pratt <mcpratt@pm.me>