kernel: of: silence bad cell count error
This warning turns out to be harmless for nodes
which are child firmware nodes of a real device
(without either a driver or hardware bus or "device" struct),
so make it a debug message.

The function is simply checking that the values
of #address-cells and #size-cells are > 0
but this is not always necessary for some nodes
in a fwnode tree that is undergoing address translation.

Signed-off-by: Michael Pratt <mcpratt@pm.me>