kernel: of: silence bad cell count error
This warning turns out to be harmless for fwnodes
which are a child of a fwnode that represents a real device
(without either a device struct or a driver struct),
so make it a debug message.

The function is simply checking that the values
of #address-cells and #size-cells are > 0
but this is not always necessary for some nodes
in devicetree that are undergoing address translation.

This error can also be silenced by adding
the #address-cells and #size-cells dt properties
as > 0 to all relevant nodes in the devicetree.

Signed-off-by: Michael Pratt <mcpratt@pm.me>