kernel: fw_devlink: link to supplier parent if no dev
Introduce flag FWNODE_FLAG_PARENT_IS_DEV
in order to mark child firmware nodes of a device
as having a parent device when that parent device
is successfully probed.

Allow fwnode link creation for suppliers that are flagged
as FWNODE_FLAG_NOT_DEVICE only if it is also flagged
as FWNODE_FLAG_PARENT_IS_DEV, which is then handled
by a new function __fwnode_link_to_parents() which creates
the fwnode link to an available parent of the fwnode instead
of the original supplier fwnode which will likely never probe.

All this depends on other functions to label a supplier fwnode
as not a real device, which must be done before the fwnode links
are created, and if after that, relevant links to the supplier
would have to be deleted and recreated, otherwise the fwnode link
would drop when device links are made, before either can be modified.

If a fwnode is flagged as FWNODE_FLAG_NOT_DEVICE
by the time a device link is to be made with it, the link is dropped.

Signed-off-by: Michael Pratt <mcpratt@pm.me>