kernel: fw_devlink: link to supplier parent if no dev
Introduce flag FWNODE_FLAG_PARENT_IS_DEV
in order to mark child firmware nodes of a device
as having a parent device that can probe.

Allow fwnode link creation to the supplier parent
when the original supplier fwnode is flagged
as both FWNODE_FLAG_NOT_DEVICE and as FWNODE_FLAG_PARENT_IS_DEV
with a new function __fwnode_link_to_parents() which creates
the fwnode link to an available parent of the fwnode.

All of this depends on other functions to label a supplier fwnode
as not a real device, which must be done before the fwnode links
are created, and if after that, relevant links to the supplier
would have to be deleted and recreated, otherwise the fwnode link
would drop when device links are made, before either can be modified.

If a fwnode is flagged as FWNODE_FLAG_NOT_DEVICE
by the time a device link is to be made with it,
but not flagged as FWNODE_FLAG_PARENT_IS_DEV,
the link is dropped.

The use case of function __fw_devlink_pickup_dangling_consumers()
is designed so that the parameters are always a supplier fwnode
and a parent of the supplier fwnode, so it is safer to assume
that the flag PARENT_IS_DEV should be added, rather than
declaring the supplier fwnode as NOT_DEVICE at that point.

Signed-off-by: Michael Pratt <mcpratt@pm.me>