kernel: fw_devlink: link to supplier parent if no device
Introduce flag FWNODE_FLAG_PARENT_IS_DEV
in order to mark child firmware nodes of a device
as having a parent device that can probe.

Allow fwnode link creation to the supplier parent
when the original supplier fwnode is flagged
as both FWNODE_FLAG_NOT_DEVICE and as FWNODE_FLAG_PARENT_IS_DEV
with a new function __fwnode_link_to_parents() which then creates
the fwnode link to a real device that provides the supplier's function.

This depends on other functions to label a supplier fwnode
as not a real device, which must be done before the fwnode links
are created, and if after that, relevant links to the supplier
would have to be deleted and recreated, otherwise the fwnode link
would be dropped before the device link is made,
and before these flags can have any effect.

If a fwnode is flagged as FWNODE_FLAG_NOT_DEVICE
by the time a device link is to be made with it,
but not flagged as FWNODE_FLAG_PARENT_IS_DEV,
the link is dropped. Also, don't use
fwnode_get_next_parent_dev() because the real device
may also be an ancestor of the next parent node.

The use case of function __fw_devlink_pickup_dangling_consumers()
is designed so that the parameters are always a supplier fwnode
and a parent of the supplier fwnode, so it is safer to assume
that the flag PARENT_IS_DEV should be added there, rather than
declaring the supplier fwnode as NOT_DEVICE at that point.

Signed-off-by: Michael Pratt <mcpratt@pm.me>