kernel: fw_devlink: retry supplier links on probe defer
When a supplier device is the cause for a consumer device
to have probe deferred and the supplier is flagged PARENT_IS_DEV,
flag the supplier as NOT_DEVICE also and recreate
all fwnode/device links to the consumer device.

Because of the flags, these links will be handled differently
when deleted and recreated compared to before the probe attempt
by linking the consumer with a parent fwnode of the original supplier,
which will later probe successfully after the consumer probe defer.

This parent of the supplier is the real device providing
the functionality that the consumer needs before probing.
(e.g. MTD partitions that are an NVMEM provider with "nvmem-cells")

In order to achieve this method, device_links_check_suppliers()
must be moved to the bottom of the function stack for device links.

Signed-off-by: Michael Pratt <mcpratt@pm.me>