ramips: kernel: set remaining subtargets to use VARIABLE_ERASE
Some devices in ramips have a u-boot environment
that is strictly in a 4k blocksize,
but setting the erasesize to 4k by default
causes a large performance loss.

Make use of the patch that allows writing in 4k blocks
while the actual blocksize is a multiple of that
in all ramips subtargets.

This is already applied in mt7260 and mt7621.

Suggested-by: Shiji Yang <yangshiji66@qq.com>
Acked-by: Thibaut VARENE <hacks@slashdirt.org>
Signed-off-by: Michael Pratt <mcpratt@pm.me>