tools/coreutils: update to 9.2
This resolves an error when building toolchain/musl on macOS due to
improper hole-detection caused by a bug in macOS/APFS:
https://github.com/openwrt/openwrt/pull/11960

As long as we don't reconfigure, 001-m4.patch is not needed.
If we keep it, it will force reconfigure the project,
since m4 files are changed. This works, but may not be optimal,
because the build should use files from coreutils/m4,
but OpenWRT uses legacy files from staging_dir/host/share/aclocal.
https://github.com/openwrt/openwrt/pull/12233#issuecomment-1481097456

Coreutils bug report and discussion
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=61386

backport a couple of upstream patches as requested by mpratt14
date: diagnose -f read errors
copy: fix --reflink=auto to fallback in more cases

Signed-off-by: Georgi Valkov <gvalkov@gmail.com>