tools: add gnulib source
By having a local copy of gnulib, we can
import the latest macro fixes into any project,
get rid of some statically stored macros that were otherwise missing,
bootstrap GNU tools with the latest relevant source
without having to wait for a release or rely on git submodules,
and possibly more.

The patch assists in bootstrapping by ignoring
the building of po files using gettext,
if that is desired, one can opt to not sync
the bootstrap script with the one in gnulib source.

Signed-off-by: Michael Pratt <mcpratt@pm.me>