tools/coreutils: bootstrap to localized gnulib source
Using a local gnulib source during bootstrap
allows for fine-grained control over the macros
and source files for use with coreutils
but published into gnulib,
without having to wait for a release
or deal with gnulib as a git submodule.

A change of source is required for this.
Using the bootstrap script that coreutils uses
does not work with a release tarball that has already had
the bootstrap step completed by maintainers,
so we now download the exact copy of source
at the commit when the release occurred.

Signed-off-by: Michael Pratt <mcpratt@pm.me>