prereq-build: do not replace binaries with symlinks
Some programs, like bash, have a symlink installed first,
and then later for some hosts, are built from source
and installed in the same directory.

Now that the prereq-build checks are not successful
just by finding the file alone, it is possible for
a new symlink to overwrite the installed binary.

If a normal file is found in staging_dir/host/bin
let the check look for the associated stampfile, and if found,
skip creation of a symlink and exit successfully.

Signed-off-by: Michael Pratt <mcpratt@pm.me>