libxml2: fix iconv dependency
libxml2 currently requires iconv during build.
Usually it will be on many Linux systems,
but not all, and usually not on macos.

Even without error, it is not desirable
for the host system's copy to be used.

Let requiring iconv to build
depend on selecting full language support
in build configuration.

Signed-off-by: Michael Pratt <mcpratt@pm.me>