tools/coreutils: update to 9.4
Update to latest stable release.

Add configure option to disable support for the Year 2038 problem.
(for now, as some versions of GCC do not yet support it)

Syncing bootstrap script fails, backport an upstream patch
which can be removed at next coreutils update.

Several headers from the stable gnulib branch cause build failure
because the changes in the imported versions are incompatible
with the Makefile that gets generated for coreutils.
This version of coreutils was released after being bootstrapped and
autoreconf'ed with a significantly different version of gnulib compared
to our local gnulib, so skip importing them (restore the backup).

While at it, organize restoring the originally shipped version
of files into a Make foreach function.

Refresh patch:
- 000-bootstrap.patch

New patch:
- 001-bootstrap-sync.patch

Link: https://lists.gnu.org/archive/html/coreutils/2023-08/msg00099.html
Tested-by: Georgi Valkov <gvalkov@gmail.com> # macOS
Signed-off-by: Michael Pratt <mcpratt@pm.me>