kernel: fw_devlink: use struct driver to determine probe ability
The function __fw_devlink_pickup_dangling_consumers()
intends to ignore suppliers that are already capable of probing,
but uses whether or not a struct bus is defined in the struct device.

There are some cases where a firmware child node
can be address translatable but not probing
(e.g. the use of of_platform_populate() for MTD partitions),
so whether or not a driver is present is a more accurate way
to guess whether a fwnode represents a real probing device here.

"Hey! Who's driving the bus?"

Signed-off-by: Michael Pratt <mcpratt@pm.me>