kernel: fw_devlink: use struct driver to determine probe ability
The dangling_consumers function intends to ignore suppliers
that are already capable of probing, but uses whether or not
a struct bus is defined. There are some cases where a firmware
child node can be address translatable but not probing,
e.g. the use of of_platform_populate() for MTD partitions,
so whether or not a driver is present is a more accurate way
to tell whether a device is probing.

Signed-off-by: Michael Pratt <mcpratt@pm.me>