kernel: of: add link exceptions to of_link_to_phandle()
Upstream commit 4a032827d ("of: property: Simplify of_link_to_phandle()")
ended up being a little too extreme of a simplification
and removed some useful logic
for fwnodes representing lower-level hardware.

Bring back the condition for not linking to "early devices" (with void return),
which will later be reinforced by additionally checking for a new flag,
and add logic to ensure that a fwnode linking to itself is ignored.

While at it, use a goto statement to save some LOC.

Signed-off-by: Michael Pratt <mcpratt@pm.me>