kernel: of: backport fix for node_not_dev
The function of_link_property() is supposed to handle cases
where the supplier is already known ahead of time to not be a probing device
(no driver or device struct) through the "compatible" property in devicetree.

Backport a commit which takes the next parent node of the supplier
with a "compatible" property as the new supplier in this case.

Signed-off-by: Michael Pratt <mcpratt@pm.me>