kernel: fw_devlink: prepare to fixup bad supplier links
The function device_links_check_suppliers() is a good place
to fixup bad supplier links, since it is a point where
the consumer is attempting a probe and the supplier
has already had a chance to probe. This is where
whether or not to defer the probe of the consumer
based on it's suppliers is decided. And, since
deferal of the probe also means deferral of
processing the devices links, it is the last opportunity
to fix the supplier before a re-probe or a different probe.

It is important to handle as many device link changes
as possible before probe deferral, since deferred_probe_initcall()
will only be called *once* at late_initcall,
and another probe defer at that point or later
can cause the consumer to be deferred indefinitely.

In order to check *every* supplier at this point
as to whether or not it needs a fixup
or to trigger a probe defer of the consumer,
use a continue statement instead of a break statement.

If a probe defer is needed by one supplier,
but immediate re-probe needed by another,
prefer the probe defer in order to satisfy
both of the supplier's requirements, so only set
the return to EAGAIN if not already set.

Signed-off-by: Michael Pratt <mcpratt@pm.me>