kernel: fw_devlink: add function device_links_fixup_suppliers()
A supplier link that points to a child firmware node of a real device
will cause the probe of the consumer device to be deferred in some cases
like when the consumer is registered as a device and attempts probe first.

When the supplier is flagged PARENT_IS_DEV
and the consumer was either already deferred or the supplier probed first,
flag the supplier as NOT_DEVICE also and recreate
all fwnode/device links to the consumer device.

Because of the flags, these links will be handled differently
when deleted and recreated compared to before the consumer probe attempt
by linking the consumer with a parent fwnode of the original supplier,
which at this point represents a device that already probed successfully.

This parent of the supplier is the real device providing
the functionality that the consumer needs before probing.
(e.g. MTD partitions that are an NVMEM provider with "nvmem-cells")

It is important to handle as many device link changes
as possible before probe deferral, since deferred_probe_initcall()
will only be called *once* at late_initcall,
and another probe defer at that point or later
can cause the consumer to be deferred indefinitely.

Handle this in a new function device_links_fixup_suppliers()
which will be called before device_links_check_suppliers()
which is where it is decided whether the consumer must be deferred.

This new function is now the last opportunity to change
device links before probe is deferred, which also stops device links
from being processed until re-probe or the next time device_add() is called.

Also, make sure that the device links are not handled earlier
than in the initial case for the fixup function, where
a device link is created to a child node of the real supplier.
Otherwise, before deferred probes are started again,
the link will be converted to sync_state only.
This method relies on the second probe to actually begin,
in the case that the deferral of the probe was for another reason.

Signed-off-by: Michael Pratt <mcpratt@pm.me>