hostapd: add basic uci support for mesh parameters
Wpad does not yet have support for iw mesh parameters
like mesh_gate_announcements and many others,
despite having support for mesh_fwding and mesh_rssi_threshold.

Unless wpad is patched (and it would likely be a huge patch),
setting the rest of the params must be handled by hotplugs
because these values will be erased anytime the interface goes down.

The new hotplug script will only be installed
when support for a mesh interface is enabled.

The script checks for bogus parameters and values
so that the while loop to check the new value is safe
and uses only the current uci configuration and the sysfs
and exit codes from iw in order to verify
that it is performing actions on the correct interface.

Linked below is a forum discussion.

Usage example:

config wifi-iface 'default_radio1'
	option device 'radio1'
	option network 'lan'
	option mode 'mesh'
	option encryption 'none'
	option mesh_id 'mesh0'
	option mesh_fwding '1'
	option mesh_rssi_threshold '-84'
	list mesh_param 'mesh_gate_announcements=1'
	list mesh_param 'mesh_connected_to_gate=1'
	list mesh_param 'mesh_connected_to_as=1'

Link: https://forum.openwrt.org/t/bug-report-802-11s-mesh-v19-07-4/78524
Signed-off-by: Michael Pratt <mcpratt@pm.me>