build: use host gzip for decompressing kernel source
libdeflate-gzip is not recommended for production use by it's authors,
and the reasons can be seen when trying to decompress a massive tarball,
like from a kernel snapshot (~ 200 MB), especially on a system with limited memory.
It is possible to get the following error:

tar: This does not look like a tar archive
tar: Exiting with failure status due to previous errors

The libdeflate version is perfectly fine for other uses,
but let's use the host system's gzip for extra large files.

Signed-off-by: Michael Pratt <mcpratt@pm.me>