tools/bison: hardcode path to m4 with STAGING_DIR_HOST
Force bison to ignore the environment variable M4
and hardcode it to the locally built m4 during build operations.

This allows bison to continue to function while we are forcefully
avoiding autoreconf and other autoconf and automake-like operations
by giving a fake path to m4 with the M4 environment variable.

The specific path can still be overridden within the line
of invocation that runs bison independent from the environment
by setting STAGING_DIR_HOST within the command,
so document this in the help printout.

Signed-off-by: Michael Pratt <mcpratt@pm.me>