tools/coreutils: update to 9.5
Update to latest stable release.

The following commits in gnulib caused a conflict
in locally bootstrapping coreutils with stable gnulib:

  8f4b4e52c991de2233b471f8e35a068866b31f01
  2749234203959df8d72cd8638d4e00a9fff450db

A module (strftime) was marked deprecated and replaced by
a module (nstrftime) in the version of gnulib that
coreutils was released with compared to the stable branch
that we use for importing. Conflicts from the previous version
of coreutils are now gone, so other imported headers are now good.

Refresh patch:
 - 000-bootstrap.patch

Remove upstreamed patch:
 - 001-bootstrap-sync.patch

Link: https://lists.gnu.org/archive/html/coreutils/2024-03/msg00132.html
Tested-by: Georgi Valkov <gvalkov@gmail.com> # macOS
Signed-off-by: Michael Pratt <mcpratt@pm.me>