tools/elfutils: fix build for Mac OS with newer version
In the latest 0.191 version of elfutils,
there is a new local source file xmalloc.c
which is significantly different than the one
that we have configured to be imported from gnulib.

The version of the file imported by gnulib calls
the function reallocarray() which is not commonly
available on Mac OS host machines.

Fix this by reverse linking the version of the file
that is released with elfutils back into
the imported gnulib subdirectory, so that only
the original version is used during build.

While at it, reorganize the Makefile lines
involved in gnulib importing and its workarounds.

Signed-off-by: Michael Pratt <mcpratt@pm.me>