tools/elfutils: override SUBDIRS variable of Makefile
Instead of editing the SUBDIRS variable with a patch,
it can be overriden at the end of the command line when invoking Make.

This tool has a series of recursive Makefiles in each subdirectory,
therefore SUBDIRS is set to a pattern of Make functions
so that the result is variable depending on the current subdirectory
that Make is being invoked in.

This is a step in the direction of no longer needing autoreconf here.

It's not necessary to have gnulib-cache.m4 in EXTRA_DIST
since we don't need to re-import after packaging this in the SDK,
so get rid of the entire patch hunk for ./Makefile.am

Signed-off-by: Michael Pratt <mcpratt@pm.me>