tools/elfutils: force linking to gnulib for reallocarray()
On macOS, stdlib.h in the standard include paths does not provide
reallocarray() while both elfutils and gnulib does, however
they are declared differently, leading to an error:

  ./system.h:101:1: error: static declaration of 'reallocarray' follows non-static declaration
  reallocarray (void *ptr, size_t nmemb, size_t size)

A normal "configure && make" build cycle results in both
declarations being enabled as a result of both elfutils
and gnulib having completely separate configure checks
where gnulib uses an internal placeholder symbol HAVE_REALLOCARRAY,
and elfutils uses a standard autoconf macro HAVE_DECL_REALLOCARRAY.

Fix this by running configure stage again after building
the local libgnu which would either contain reallocarray()
when missing and then be available to be linked for it afterwards,
or not contain reallocarray() and trigger elfutils local copy
to be used instead.

Having a local include with GNU's stdlib.h is safe in this case
because the header uses the include_next directive in order
to include the standard stdlib.h include afterwards,
and it doesn't conflict with those standard headers.

Signed-off-by: Michael Pratt <mcpratt@pm.me>