tools/elfutils: use locally declared static reallocarray()
For the new 0.191 version of elfutils, importing the dirname
module from gnulib is needed, and this adds a dependency for
the reallocarray module which provides the reallocarray() function.

On macOS, stdlib.h in the standard include paths does not provide
reallocarray() while both elfutils and gnulib does, however
they are declared differently, leading to an error:

  ./system.h:101:1: error: static declaration of 'reallocarray' follows non-static declaration
  reallocarray (void *ptr, size_t nmemb, size_t size)

A normal "configure && make" build cycle results in both
declarations being enabled as a result of both elfutils
and gnulib having completely separate configure checks
where gnulib uses an internal placeholder symbol HAVE_REALLOCARRAY,
and elfutils uses a standard autoconf macro HAVE_DECL_REALLOCARRAY.

Fix this by excluding the import of the reallocarray module ahead of time
which causes gnulib checks in the configure stage to not even consider
whether to declare reallocarray later on, so the decision is
only between the standard include stdlib.h and the elfutils header.

Tested-by: Georgi Valkov <gvalkov@gmail.com> # macOS, elfutils 0.191
Signed-off-by: Michael Pratt <mcpratt@pm.me>