tools/gnulib: add header and conditonals for tdestroy()
The tdestroy() function, which is a GNU extension
to the standard C library, is defined in gnulib
in tsearch.c but is missing it's corresponding header
in search.in.h by being completely missing...

This patch adds it, although it is not upstream friendly yet.
Upstream GNU would want to see a configure check that looks
for whether tdestroy() is declared first, and use the @@
placeholder/substitution method for whether to declare it.

Signed-off-by: Michael Pratt <mcpratt@pm.me>