tools/gnulib: make tdestroy() fully portable
The tdestroy() function, which is a GNU extension
to the standard C library, is defined in gnulib
in tsearch.c but is missing it's corresponding declaration
in search.in.h by being completely missing...

This patch is large but upstreamable, including all
of the macros and conditionals and configure checks
that upstream GNU would expect for portable support,
like using the @@ placeholder/substitution method
to determine whether or not to have declarations
based on whether or not tdestroy() is already declared
within the standard headers of the default include paths.

There was also some typedefs and aliases missing,
along with the warnings and preprocessor exceptions
that need to be added for consistency
with the usage of the rest of the functions in the files.

Signed-off-by: Michael Pratt <mcpratt@pm.me>