tools/elfutils: remove no-op copy of tdestroy()
A false tdestroy() function was added in order to make elfutils
build on macOS again. The previous commit adds declarations for a
real version of tdestroy() into gnulib, which is already imported,
as well as the preprocessor flags and the triggers for the Makefile.am
conditional in order to include the source to be built.

Tested-by: Georgi Valkov <gvalkov@gmail.com> # macOS
Signed-off-by: Michael Pratt <mcpratt@pm.me>