tools/elfutils: refresh portability patch for macOS
Quilt refresh combined two sets of changes to the same file.

The switch from using libgen.h to dirname.h because of
function poisoning from gnulib's import of basename() was added
as a new patch hunk instead of an edit to the original one.

The original patch hunk was to fix build errors
on an earlier version of elfutils
before the "dirname" module was being imported
to fix further build errors with the 0.191 version.

Signed-off-by: Michael Pratt <mcpratt@pm.me>