tools/gnulib: add macros to skip reallocarray() functions
For modules that depend on the reallocarray module,
like ialloc, xalloc, and safe-alloc,
it was not possible to skip importing the reallocarray module
as they all contained at least one function that called reallocarray()
and would cause build failure if the host system didn't declare it.

This upstreamable patch adds macros that toggle whether to define
functions that depend on reallocarray() based on whether the
reallocarray module is being imported.

Signed-off-by: Michael Pratt <mcpratt@pm.me>