tools/missing-macros: add symlinks to makeinfo
There are other wrapper scripts released with makeinfo
like texi2pdf which are required by the build
prerequisites of some tools, and have a similar
purpose and usage.

Let the makeinfo perl script handle all of these cases.

It's worth mentioning that "texi2any" is the actual program
and "makeinfo" is one of it's aliases. From upstream GNU:

  makeinfo: texi2any
	rm -f $@
	-$(LN_S) texi2any $@

Signed-off-by: Michael Pratt <mcpratt@pm.me>