tools/elfutils: do not link gnulib to libelf
The compiled library resulting from importing gnulib to the build
has been linked to libelf in order to easily cover linking dependencies
to each of the binaries. However, this is not appropriate
as it bloats the final libelf library, and may result in
multiple defintions of symbols based on whether or not
certain modules from gnulib are included. This is not
a problem while building elfutils, because gnulib has it's own
special way of creating function aliases and special declarations
that allow the linker to ignore the original function definitions,
however, libelf may be used to link to other things, resulting in
an error at link time.

Fix this by removing the linking between gnulib and libelf
and instead overriding Make variables in order to add linking
between gnulib and each of the binaries provided by elfutils.

Signed-off-by: Michael Pratt <mcpratt@pm.me>