tools/elfutils: do not link gnulib to libelf
The compiled library resulting from importing gnulib has been
linked to libelf in order to easily cover other link dependencies.
However, this is not appropriate for linking to other programs
as it bloats the final libelf library, and may result in
multiple defintions of symbols based on whether or not
certain modules from gnulib are included.

This is not a problem while building elfutils, because gnulib has
it's own way of creating function aliases and special declarations
that allow the linker to ignore the original function definitions,
however, when libelf is used to link to something else,
this results in an error at link time.

Fix this by removing the linking between gnulib and libelf
and instead overriding Make variables in order to add linking
between gnulib and each of the binaries provided by elfutils,
using Make functions to avoid applying it to other subdirectories.

Block unnecessary replacement of some functions by gnulib
so that future linking with libelf doesn't require
the associated libgnu library from this build.

Signed-off-by: Michael Pratt <mcpratt@pm.me>