tools/gnulib: unmangle fts header
The gnulib fts.h header is meant to not be overwritten
in any way by the host system's copy of fts.h
and is therefore given a unique name instead.

This is fine if the built libgnu library is directly linked
with the target library, but if we want to keep them isolated
we end up having the definitions being mangled anyway
when the next object to link against included the fts.h header.

On some macOS platforms, the use of __DARWIN_INODE64
is messing with the link name for fts functions, resulting in:

Undefined symbols for architecture x86_64:
  "_rpl_fts_close$INODE64", referenced from:
...

The file was named with an underscore for historical reasons
before the include guard was changed
to match the standard for Linux systems.

Rename the header back to fts.h
and add the non-standard include guard for macOS.

Signed-off-by: Michael Pratt <mcpratt@pm.me>