tools/libtool: add maintainer-clean to clean target
Use the clean recipe during build, so that files already
generated by the bootstrapping within the release are removed
before the bootstrap script is run again.

Automake with modified rules must be ran
in order to generate a Makefile that does delete itself.

A patch is added handling a bug where if the variable is blank,
then the corresponding `rm` command fails.

Signed-off-by: Michael Pratt <mcpratt@pm.me>