tools/elfutils: add maintainer-clean to clean target
Use the clean recipe during build, so that files already
generated by the bootstrapping within the release are removed
before the bootstrap script is run again.

Automake with modified rules must be ran
in order to generate a Makefile that does delete itself or libtool.

In order to enable rules for building what is removed,
maintainer mode is required for configuring.

Building from tracked sources only with maintainer mode
requires a lexical analyzer, so add dependency for bison.

Add a header issue related to use of gnulib with the lex source to patch.

The function fputc_unlocked() is not present on macOS,
so use fprintf instead in libcpu.

The function obstack_printf() is not present on macOS,
so use the available gnulib module and their dependencies.

The variables for generic automake clean rules
should not be defined conditionally, so move it
to the end of the automake file with a patch.

Signed-off-by: Michael Pratt <mcpratt@pm.me>